renaming the exec and spawn wrapper file to more correctly reflect the content
